Moved FakeFileActivityEvent to own file
