+++ sensor/kubernetes/fake/fileactivity.go
@@ -8,7 +8,7 @@ import (
 
 	sensorAPI "github.com/stackrox/rox/generated/internalapi/sensor"
 	"github.com/stackrox/rox/pkg/uuid"
-	"github.com/stackrox/rox/sensor/common/pubsub"
+	"github.com/stackrox/rox/sensor/common/events"
 	"google.golang.org/protobuf/types/known/timestamppb"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 )
@@ -65,7 +65,7 @@ func (w *WorkloadManager) manageFileActivity(ctx context.Context) {
 				continue
 			}
 
-			event := &FakeFileActivityEvent{
+			event := &events.FakeFileActivityEvent{
 				Activity: activity,
 			}
 
@@ -204,19 +204,4 @@ func generateFileActivityPaths(n int) []string {
 		paths = append(paths, fmt.Sprintf("%s/file-%s.conf", dir, uuid.NewV4().String()[:8]))
 	}
 	return paths
-}
-
-// FakeFileActivityEvent wraps a file activity message from fake workloads.
-// This allows fake workloads to publish file activities through the pubsub system rather
-// than writing directly to a shared channel.
-type FakeFileActivityEvent struct {
-	Activity *sensorAPI.FileActivity
-}
-
-func (e *FakeFileActivityEvent) Topic() pubsub.Topic {
-	return pubsub.FakeFileActivityTopic
-}
-
-func (e *FakeFileActivityEvent) Lane() pubsub.LaneID {
-	return pubsub.FakeFileActivityLane
 }
